rewriting video writing
Not sure when last commit was but the old video writer and display crashed most of the time (among other issues)
has been completely rewritten to be more synchronous in testwriter and VideoTester.
It has a memory issue again so it crashes on longer videos, but it does work very well outside of that.
Also the body pose detection analyzer kinda works but hasn't been re-implemented yet